% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuclideanDistance.R
\name{DIST}
\alias{DIST}
\title{A Euclidean Distance Finder}
\usage{
DIST(data.mat, digits = 15)
}
\arguments{
\item{data.mat}{\itemize{
\item An N x p matrix of data.
}}

\item{digits}{\itemize{
\item The number of significant digits to return.
}}
}
\value{
distance.mat - An N x N pairwise distance matrix.
}
\description{
This algorithm uses the Gram matrix to calculate
the pairwise Euclidean distance between the N
elements of data.mat. In this case, it will often
be used as a dissimilarity matrix.
}
