% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadialKernelConverter.R
\name{RADIAL_KERNEL}
\alias{RADIAL_KERNEL}
\title{Radial Kernel Generator}
\usage{
RADIAL_KERNEL(euclidian_distance.mat, c = 1)
}
\arguments{
\item{euclidian_distance.mat}{\itemize{
\item A Eclidean distance matrix
}}

\item{c}{\itemize{
\item A scale factor
}}
}
\value{
RMG.mat
}
\description{
Converts a Euclidean Distance (dissimilarity) matrix into a
similarity matrix for spectral analysis.
}
\examples{
similarity.mat <- RADIAL_KERNEL(as.matrix(DIST(iris[,1:4])))
similarity.mat[1:10,1:10]
}
