% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECC_RK.R
\name{SPECC_RK}
\alias{SPECC_RK}
\title{Spectral Clustering using a Radial Kernel}
\usage{
SPECC_RK(data.mat, centers, scale = 1)
}
\arguments{
\item{data.mat}{\itemize{
\item An N x p data matrix.
}}

\item{centers}{\itemize{
\item The number of cluster centers returned.
}}

\item{scale}{\itemize{
\item A scale factor for the radial kernel function.
}}
}
\value{
kmeans output - TODO update this
}
\description{
Spectral Clustering using a Radial Kernel
}
