% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadialKernelConverter.R
\name{RADIAL_KERNEL}
\alias{RADIAL_KERNEL}
\title{Radial Kernel Generator}
\usage{
RADIAL_KERNEL(euclidian_distance.mat, c = 1)
}
\arguments{
\item{c}{}
}
\value{
RMG.mat
}
\description{
Converts a Euclidean Distance (dissimilarity) matrix into a
similarity matrix for spectral analysis.
}
\examples{
RADIAL_KERNEL(dist(iris[,1:4]))
}
